Reconcile the caisse when a paid order is re-priced, and stop double-counting caisse-settled documents
Two defects, both surfacing as the marge nette moving on its own.

Reproduced first. An order of 1000 HT / 1200 TTC settled from the caisse,
then re-priced 100 -> 50: order HT fell 500, caisse debit stayed 1200,
and the marge nette rose 500. Money never moved, yet the margin improved,
because the order and the cash ledger were reporting different amounts.

updateLignePrix now posts the difference back to the caisse when the
order is PAYE and was settled by CAISSE — as its own movement, so the
ledger stays append-only and the correction is visible as a correction.
Other payment modes clear outside the caisse and are left alone. Same
reproduction after the fix: marge nette moves 0.00, and the caisse shows
DEBIT 1200 followed by CREDIT 600.

The second defect was underneath it. decaissementsGlobaux added caisse
debits to settled achats, sous-traitance and salaries, counting a
caisse-settled document twice. The comment said this was an
approximation because payment source was not tracked — it is now, so the
sums exclude what the caisse ledger already contains.

Also widens lignes_achat.quantite to DOUBLE PRECISION (migration 036), so
both sides of a line total carry the same precision. DECIMAL(15,3) capped
quantities at three decimals while the price beside it had none. Line
totals stay DECIMAL(15,2). Stock quantities keep their scale and are
rounded HALF_UP on the way in rather than truncated silently.

+++ src/main/java/com/buildflow/erp/domain/achats/dto/response/LigneAchatResponse.java
@@ -7,7 +7,7 @@ public record LigneAchatResponse(
         UUID id,
         String articleCode,
         String designation,
-        BigDecimal quantite,
+        double quantite,
         String unite,
         double prixUnitaire,
         BigDecimal total,
